myproject/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.4
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.27.2
pyarrow==16.1.0
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==5.4.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.37.2
streamlit==1.36.0
systemd-python==234
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wadllib==1.3.6
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
zipp==1.0.0
